add table + todo list

+++ backend/app/models/project.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Boolean, DateTime, Text, Float
+from sqlalchemy import Column, Integer, String, Boolean, DateTime, Text, Float, BigInteger
 from sqlalchemy.types import JSON
 from app.core.database import Base
 
@@ -19,9 +19,9 @@ class Project(Base):
     implem_end = Column(DateTime)
     winner = Column(Boolean, index=True)
     rate = Column(Float)
-    money_req_grant = Column(Integer)
-    cofunding = Column(Integer)
-    total_money = Column(Integer)
+    money_req_grant = Column(BigInteger)  # Исправлено для больших чисел
+    cofunding = Column(BigInteger)        # Исправлено для больших чисел  
+    total_money = Column(BigInteger)      # Исправлено для больших чисел
     description = Column(Text)
     goal = Column(Text)
     tasks = Column(Text)
